+++ src/components/fields/Field_SingleRelationship.vue
@@ -613,7 +613,7 @@ export default class Field_SingleRelationship extends FieldBase {
       const needle = val.toLowerCase()
       this.filterList = extend(true, [], this.allTypeDocuments)
       // @ts-ignore
-      this.filterList = advancedDocumentFilter(needle, this.filterList, this.SGET_allBlueprints, this.filterList)
+      this.filterList = advancedDocumentFilter(needle, this.filterList, this.SGET_allBlueprints, this.SGET_allDocuments.docs)
         .filter((obj) => obj._id !== this.currentId)
 
       if (this.$refs[`singleRelationshipField${this.inputDataBluePrint.id}`] && this.filterList.length > 0) {